chore: Move logger to core/logger
