+++ src/commands/publish/run.ts
@@ -1,5 +1,5 @@
 import type {PublishConfig} from './index';
-import {Logger} from '~/logger';
+import {Logger} from '~/core/logger';
 import {PutObjectCommand, S3Client} from '@aws-sdk/client-s3';
 import {normalizePath} from '~/utils';
 import {join, resolve} from 'path';
