+++ src/commands/publish/upload.ts
@@ -3,7 +3,7 @@ import {join} from 'path';
 import {asyncify, mapLimit} from 'async';
 import walkSync from 'walk-sync';
 import mime from 'mime-types';
-import {LogLevel} from '~/logger';
+import {LogLevel} from '~/core/logger';
 
 export async function upload(run: Run): Promise<void> {
     const {input, endpoint, bucket, prefix, hidden = []} = run.config;
